pwm: Drop changes related to pwm_config() compared to mainline
pwm_config() isn't recommended to be used any more and (since the
previous commit) the vendor changes don't introduce any new user. So
drop the changes introduced in commit 2a1b16e2639a ("pwm: Convert period
and duty cycle to u64").

(cherry picked from commit 526af93b8f12021ea86438539c381bdc611d9183)

+++ include/linux/pwm.h
@@ -323,8 +323,8 @@ int pwm_adjust_config(struct pwm_device *pwm);
  *
  * Returns: 0 on success or a negative error code on failure.
  */
-static inline int pwm_config(struct pwm_device *pwm, u64 duty_ns,
-			     u64 period_ns)
+static inline int pwm_config(struct pwm_device *pwm, int duty_ns,
+			     int period_ns)
 {
 	struct pwm_state state;
 
@@ -423,8 +423,8 @@ static inline int pwm_adjust_config(struct pwm_device *pwm)
 	return -ENOTSUPP;
 }
 
-static inline int pwm_config(struct pwm_device *pwm, u64 duty_ns,
-			     u64 period_ns)
+static inline int pwm_config(struct pwm_device *pwm, int duty_ns,
+			     int period_ns)
 {
 	might_sleep();
 	return -EINVAL;
